refactor: remove WritableVecA
